% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_himawari.R
\name{geom_himawari}
\alias{geom_himawari}
\alias{GeomHimawari}
\title{Geom objects for himawari plots}
\usage{
geom_himawari(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}
}
\description{
This plot is designed for the situation where each datapoint is defined by a unique combination of two variables and is associated with two statistics, the second of which is bounded in [0, 1].
Each datapoint is visualized by a sunflower (hence, himawari) where the first statistic is used for the color and the second statistic is used to determine the number of petals.
}
\section{Motivation}{

Our aim is to avoid using the size to visualize the second statistic.
This complicates interpretation as the background color effectively becomes a new endpoint on the color scale,
requiring readers to consider two dimensions of color transitions.
At worst, the background color is the same as one of the existing endpoints and the interpretation is confounded.

By comparison, the himawari plot is simpler as the color scale only has one dimension that is reserved for the first statistic.
At a glance, the plot can be easily interpreted in the same manner as a heatmap.
The second statistic is deliberately less visible to simplify interpretation,
but is still available for a more detailed inspection of the plot.

The \code{\link{sunflowerplot}} name was already taken, hence the unusual name for this function.
At least it sounds artistic.
}

\section{Aesthetics}{

\code{geom_himawari} requires the following aesthetics:
\itemize{
\item \code{x}, a categorical variable used to determine the x-axis position of the sunflower for each row in \code{data}.
\item \code{y}, a categorical variable used to determine the y-axis position of the sunflower for each row in \code{data}.
\item \code{fill}, a numeric variable used to determine the color of each sunflower for each row in \code{data}.
\item \code{percent}, a numeric variable containing values from [0, 1], 
used to determine the number of petals on the sunflower for each row in \code{data}.
}
It also understands the following optional aesthetics:
\itemize{
\item \code{color}, the color of the petals.
Usually set to the same value as \code{fill}.
\item \code{ratio}, the ratio of the radius of the sunflower's face to the petal length.
\item \code{size}, the width of the lines used to draw the petals.
\item \code{npetals}, the total number of petals on a sunflower with \code{percent=1}.
}
}

\examples{
stuff <- data.frame(X=rep(LETTERS, 5), 
   Y=as.character(rep(1:5, each=26)), 
   LogFC=rnorm(26*5), 
   Prop=runif(26*5))

ggplot(stuff) + geom_himawari(aes(x=X, y=Y, fill=LogFC, percent=Prop))

}
