% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHimawari.R
\name{plotHimawari}
\alias{plotHimawari}
\title{Create a himawari plot}
\usage{
plotHimawari(logfc, percent, colors = TRUE, ...)
}
\arguments{
\item{logfc}{A numeric matrix of statistics, usually a log-fold change for each gene (row) in each cluster (column)
compared to the average across all clusters.}

\item{percent}{A numeric matrix with the same dimensions as \code{logfc}, containing the percentage of expressing cells.}

\item{colors}{A logical scalar indicating whether a blue-white-red color scheme should be added automatically.
This can be \code{FALSE} to allow colors to be manually set if values other than log-FCs are used in \code{logfc}.}

\item{...}{Arguments to pass to \code{\link{geom_himawari}}.}
}
\value{
A \link{ggplot} object containing the himawari plot.
}
\description{
Create a himawari (sunflower) plot to simultaneously visualize log-fold changes and the percentage of expressing cells.
}
\details{
Each matrix entry is represented by a sunflower in a grid.
\code{logfc} is used for coloring of the sunflower, while \code{percent} is used to determine the number of petals.
The motivation for this approach is discussed in more detail in \code{?\link{geom_himawari}}.
}
\examples{
############################
#### Setting up the data ###
############################

library(scRNAseq)
sce <- ZeiselBrainData()

library(scran)
library(scater)
sce <- logNormCounts(sce)
markers <- pairwiseTTests(logcounts(sce), sce$level1class)
output <- getTopMarkers(markers[[1]], markers[[2]], n=2)

features <- unlist(unlist(output))
object <- sce
group <- sce$level1class

##################################
#### Setting up the statistics ###
##################################

num <- numDetectedAcrossCells(object, ids = group, 
    subset_row = features, average = TRUE)
ave <- sumCountsAcrossCells(object, ids = group, subset_row = features,
    exprs_values="logcounts", average = TRUE)

ave <- ave - rowMeans(ave)

########################
#### Making the plot ###
########################

plotHimawari(ave, num) 

}
\seealso{
\code{\link{geom_himawari}}, for more manual control over the himawari parameters.
}
\author{
Aaron Lun
}
