% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHimawari.R
\name{plotHimawariBase}
\alias{plotHimawariBase}
\title{Create a himawari plot}
\usage{
plotHimawariBase(stat1, stat2, radius = 0.3, npetals = 10,
  petal.len = 0.1, petal.width = 2)
}
\arguments{
\item{stat1}{A numeric matrix of statistics.}

\item{stat2}{A numeric matrix with the same dimensions as \code{stat1}, but with values in [0, 1].}

\item{radius}{Numeric scalar specifying the size of the sunflower faces.
Should be less than 0.5.}

\item{npetals}{Integer scalar specifying the maximum number of petals.}

\item{petal.len}{Numeric scalar specifying the length of the petals.
Should be less than \code{0.5 - radius}.}

\item{petal.width}{Numeric scalar specifying the width of the petals.}
}
\value{
A \link{ggplot} object containing the himawari plot.
}
\description{
Create a himawari (sunflower) plot to simultaneously visualize two matrices of statistics.
Each matrix entry is represented by a sunflower in a grid.
The primary matrix is used for coloring of the sunflower, while the secondary matrix is used to determine the number of petals.
}
\details{
This plot aims to visualize two matrices in a single grid using sunflowers to represent each entry of the matrix.
The first matrix is visualized using color, and the second matrix is visualized based on the number of petals.
As such, values of the second matrix are expected to lie within [0, 1].

The motivation is to avoid using the size to visualize \code{stat2}.
This complicates interpretation as the background color effectively becomes a new endpoint on the color scale,
requiring readers to consider two dimensions of color transitions.
At worst, the background color is the same as one of the existing endpoints and the interpretation is confounded.

By comparison, the himawari plot is simpler as the color scale only has one dimension for \code{stat1}.
At a glance, the plot can be easily interpreted in the same manner as a heatmap.
\code{stat2} is pushed into the background but is still available for a more detailed inspection of the plot.

The \code{\link{sunflowerplot}} name was already taken, hence the unusual name for this function.
At least it sounds artistic.
}
\examples{
############################
#### Setting up the data ###
############################

library(scRNAseq)
sce <- ZeiselBrainData()

library(scran)
library(scater)
sce <- logNormCounts(sce)
markers <- pairwiseTTests(logcounts(sce), sce$level1class)
output <- getTopMarkers(markers[[1]], markers[[2]], n=2)

features <- unlist(unlist(output))
object <- sce
group <- sce$level1class

##################################
#### Setting up the statistics ###
##################################

num <- numDetectedAcrossCells(object, ids = group, 
    subset_row = features, average = TRUE)
ave <- sumCountsAcrossCells(object, ids = group, subset_row = features,
    exprs_values="logcounts", average = TRUE)

ave <- ave - rowMeans(ave)

########################
#### Making the plot ###
########################

# Adding a standard red-white-blue color scheme
G <- plotHimawariBase(ave, num) + 
    scale_fill_gradient2(low="blue", high="red") +
    scale_color_gradient2(low="blue", high="red", mid="grey", guide=FALSE)
}
\author{
Aaron Lun
}
